# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: TRC Control Plane API
  version: "v1"
  description: |
    Control plane HTTP API.

    Key properties:
    - Org-scoped paths for tenant resources.
    - Append-only event log as source of truth, materialized views for reads.
    - Workloads run as Firecracker microVM instances on hosts.
    - Ingress is L4-first (TLS SNI passthrough by default).
servers:
  - url: /v1
...
security:
  - bearerAuth: []
paths:
...
  /orgs/{org_id}/apps/{app_id}/releases:
    get:
      tags: [Releases]
      summary: List releases
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Cursor"
      responses:
        "200":
          description: Releases
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListReleasesResponse"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
    post:
      tags: [Releases]
      summary: Create release (image ref + manifest)
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateReleaseRequest"
      responses:
        "200":
          description: Release created
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "400":
          $ref: "#/components/responses/Error400"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "409":
          $ref: "#/components/responses/Error409"
  /orgs/{org_id}/apps/{app_id}/releases/{release_id}:
    get:
      tags: [Releases]
      summary: Get release
      parameters:
        - $ref: "#/components/parameters/OrgId"
        - $ref: "#/components/parameters/AppId"
        - $ref: "#/components/parameters/ReleaseId"
      responses:
        "200":
          description: Release
          headers:
            X-Request-Id:
              $ref: "#/components/headers/XRequestId"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Release"
        "401":
          $ref: "#/components/responses/Error401"
        "403":
          $ref: "#/components/responses/Error403"
        "404":
          $ref: "#/components/responses/Error404"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  headers:
    XRequestId:
      description: Request correlation id
      schema:
        type: string
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Opaque idempotency key for retry-safe writes
      schema:
        type: string
        minLength: 8
        maxLength: 128
    OrgId:
      name: org_id
      in: path
      required: true
      schema:
        type: string
...
    AppId:
      name: app_id
      in: path
      required: true
      schema:
        type: string
...
    ReleaseId:
      name: release_id
      in: path
      required: true
      schema:
        type: string
...
    Cursor:
      name: cursor
      in: query
      required: false
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
...
  responses:
    Error400:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error401:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error403:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error404:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Error409:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id, retryable]
      properties:
        code:
          type: string
        message:
          type: string
        request_id:
          type: string
        retryable:
          type: boolean
        details:
          type: object
          additionalProperties: true
...
    Release:
      type: object
      required:
        [
          id,
          org_id,
          app_id,
          image_ref,
          image_digest,
          manifest_schema_version,
          manifest_hash,
          command,
          resource_version,
          created_at,
        ]
      properties:
        id:
          type: string
        org_id:
          type: string
        app_id:
          type: string
        image_ref:
          type: string
        image_digest:
          type: string
        manifest_hash:
          type: string
        manifest_schema_version:
          type: integer
        command:
          type: array
          items:
            type: string
          minItems: 1
        resource_version:
          type: integer
        created_at:
          type: string
...
    CreateReleaseRequest:
      type: object
      required: [image_ref, image_digest, manifest_hash, command]
      properties:
        image_ref:
          type: string
          description: OCI reference (must be pinned by digest in v1)
        image_digest:
          type: string
          description: Image digest (sha256:...)
        manifest_schema_version:
          type: integer
          default: 1
        manifest_hash:
          type: string
          description: Hash of manifest content (sha256:...)
        command:
          type: array
          items:
            type: string
          minItems: 1
          description: Fully resolved entrypoint command
    ListReleasesResponse:
      type: object
      required: [items, next_cursor]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Release"
        next_cursor:
          type: [string, "null"]
...